rename corresponding pmob file
